+++ object/context.go
@@ -1,4 +1,4 @@
-package spec
+package object
 
 import (
 	"encoding/json"
